+++ src/components/qna/CategoryFilter/index.ts
@@ -1,0 +1,1 @@
+export { CategoryFilter } from './CategoryFilter'
