+++ src/components/qna/QuestionCard/index.ts
@@ -1,0 +1,1 @@
+export { QuestionCard } from './QuestionCard'
