+++ src/components/qna/index.ts
@@ -1,2 +1,4 @@
 export * from './MarkdownEditor'
 export * from './AnswerForm'
+export * from './QuestionCard'
+export * from './CategoryFilter'
